feat: wip fix date, use config
