+++ Cargo.toml
@@ -5,10 +5,13 @@ edition = "2021"
 
 
 [dependencies]
+chrono = "0.4.38"
 color-eyre = "0.6.2"
 eyre = "0.6.12"
 itertools = "0.13.0"
 lazy_static = "1.4.0"
 reqwest = { version = "0.12", features = ["blocking"] }
 scraper = "0.19.0"
+serde = { version = "1.0.203", features = ["derive"] }
+serde_json = "1.0.117"
 substring = "1.4.5"
