--- src/teachings.rs
@@ -1,1 +1,0 @@
-mod teachings;
